Refactor travel story components: replace ViewTravelStory with a new implementation in User directory and update routing in App.jsx
